Updated indic_tokenize.py
Tokenizes  the Devnagri and Gujrati numbers which appear as comma separated values in amounts.
Takes care of dates with `-`
Also, it takes care of numbers which appear at the beginning of the string. The earlier code was splitting the numbers , dates when they 
appear at the beginning of the sentence on a newline.